Move SoundpacksTab out of ui.py
