+++ cddagl/ui/views/dialogs.py
@@ -1,0 +1,137 @@
+import html
+import logging
+import os
+
+from PyQt5.QtCore import Qt
+from PyQt5.QtWidgets import (
+    QWidget, QGridLayout, QLabel, QLineEdit, QPushButton, QFileDialog, QToolButton,
+    QDialog, QTextBrowser, QMessageBox, QHBoxLayout, QTextEdit
+)
+
+import cddagl
+from cddagl.functions import clean_qt_path
+from cddagl.i18n import proxy_gettext as _
+from cddagl.win32 import get_downloads_directory
+
+version = cddagl.__version__
+logger = logging.getLogger('cddagl')
+
+
+class BrowserDownloadDialog(QDialog):
+    def __init__(self, name, url, expected_filename):
+        super(BrowserDownloadDialog, self).__init__()
+
+        self.name = name
+        self.url = url
+        self.expected_filename = expected_filename
+        self.downloaded_path = None
+
+        layout = QGridLayout()
+
+        info_label = QLabel()
+        info_label.setText(_('This {name} cannot be directly downloaded by the '
+            'launcher. You have to use your browser to download it.').format(
+            name=name))
+        layout.addWidget(info_label, 0, 0, 1, 2)
+        self.info_label = info_label
+
+        step1_label = QLabel()
+        step1_label.setText(_('1. Open the URL in your browser.'))
+        layout.addWidget(step1_label, 1, 0, 1, 2)
+        self.step1_label = step1_label
+
+        url_tb = QTextBrowser()
+        url_tb.setText('<a href="{url}">{url}</a>'.format(url=html.escape(url)))
+        url_tb.setReadOnly(True)
+        url_tb.setOpenExternalLinks(True)
+        url_tb.setMaximumHeight(23)
+        url_tb.setLineWrapMode(QTextEdit.NoWrap)
+        url_tb.setHorizontalScrollBarPolicy(Qt.ScrollBarAlwaysOff)
+        layout.addWidget(url_tb, 2, 0, 1, 2)
+        self.url_tb = url_tb
+
+        step2_label = QLabel()
+        step2_label.setText(_('2. Download the {name} on that page and wait '
+            'for the download to complete.').format(name=name))
+        layout.addWidget(step2_label, 3, 0, 1, 2)
+        self.step2_label = step2_label
+
+        step3_label = QLabel()
+        step3_label.setText(_('3. Select the downloaded archive.'))
+        layout.addWidget(step3_label, 4, 0, 1, 2)
+        self.step3_label = step3_label
+
+        path = get_downloads_directory()
+        if expected_filename is not None:
+            path = os.path.join(path, expected_filename)
+
+        download_path_le = QLineEdit()
+        download_path_le.setText(path)
+        layout.addWidget(download_path_le, 5, 0)
+        self.download_path_le = download_path_le
+
+        download_path_button = QToolButton()
+        download_path_button.setText('...')
+        download_path_button.clicked.connect(self.set_download_path)
+        layout.addWidget(download_path_button, 5, 1)
+        self.download_path_button = download_path_button
+
+        buttons_container = QWidget()
+        buttons_layout = QHBoxLayout()
+        buttons_layout.setContentsMargins(0, 0, 0, 0)
+        buttons_container.setLayout(buttons_layout)
+
+        install_button = QPushButton()
+        install_button.setText(_('Install this {name}').format(name=name))
+        install_button.clicked.connect(self.install_clicked)
+        buttons_layout.addWidget(install_button)
+        self.install_button = install_button
+
+        do_not_install_button = QPushButton()
+        do_not_install_button.setText(_('Do not install'))
+        do_not_install_button.clicked.connect(self.do_not_install_clicked)
+        buttons_layout.addWidget(do_not_install_button)
+        self.do_not_install_button = do_not_install_button
+
+        layout.addWidget(buttons_container, 6, 0, 1, 2, Qt.AlignRight)
+        self.buttons_container = buttons_container
+        self.buttons_layout = buttons_layout
+
+        self.setLayout(layout)
+
+        self.setWindowTitle(_('Browser download'))
+
+    def set_download_path(self):
+        options = QFileDialog.DontResolveSymlinks
+        selected_file, selected_filter = QFileDialog.getOpenFileName(self,
+            _('Downloaded archive'), self.download_path_le.text(),
+            _('Archive files {formats}').format(formats='(*.zip *.rar *.7z)'),
+            options=options)
+        if selected_file:
+            self.download_path_le.setText(clean_qt_path(selected_file))
+
+    def install_clicked(self):
+        choosen_file = self.download_path_le.text()
+        if not os.path.isfile(choosen_file):
+            filenotfound_msgbox = QMessageBox()
+            filenotfound_msgbox.setWindowTitle(_('File not found'))
+
+            text = (_('{filepath} is not an existing file on your system. '
+                'Make sure to download the archive with your browser. Make '
+                'sure to select the downloaded archive afterwards.')).format(
+                    filepath=choosen_file
+                )
+
+            filenotfound_msgbox.setText(text)
+            filenotfound_msgbox.addButton(_('I will try again'),
+                QMessageBox.AcceptRole)
+            filenotfound_msgbox.setIcon(QMessageBox.Warning)
+            filenotfound_msgbox.exec()
+
+            return
+
+        self.downloaded_path = choosen_file
+        self.done(1)
+
+    def do_not_install_clicked(self):
+        self.done(0)
